pyproject.toml
[project]
name = "text2sql-rocketlab"
version = "0.1.0"
description = "Agente Text-to-SQL para atividade Rocket Lab 2026"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai-slim[google]>=1.81.0",
    "python-dotenv>=1.0.1",
    "annotated-types>=0.7.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/text2sql"]
